+++ NEWS.md
@@ -3,6 +3,7 @@
   
 ## New features
  
+- New arguments `bxp.errorbar` added to `ggboxplot()` for adding error bars at the top of the box plots ([@	j3ypi, #105](https://github.com/kassambara/ggpubr/issues/105).
 - New function `stat_pvalue_manual()` for adding p-values generated elswhere ([@achamess, #81](https://github.com/kassambara/ggpubr/issues/81), [@grst, #65](https://github.com/kassambara/ggpubr/issues/65)).