+++ NEWS.md
@@ -1,3 +1,15 @@
+# ggpubr 0.1.8.999
+  
+## New features
+
+
+## Major changes
+
+
+## Minor changes
+
+
+
 # ggpubr 0.1.8